corrected security failure in cli scripts. Updated software-get to allow autoinstall.

+++ multiseat.php
@@ -1,3 +1,4 @@
 <?php
+if (!isset($_SERVER['HOME'])){die();}
 unlink('./singleseat.conf.php');
 ?>
